requirements.txt
flask
flask-login
openai
python-dotenv
PyPDF2
gunicorn
markdown